Fixing inconsistency in types in a function declartion and definition.
Former-commit-id: d5b6d3abfac2ef1305f4a9a3b3853b5dd0292f64

+++ include/hpx/lco.h
@@ -253,7 +253,7 @@ hpx_status_t hpx_lco_sema_p(hpx_addr_t sema);
 /// @param inputs the number of inputs to the and (must be >= 0)
 ///
 /// @returns The global address of the new and gate.
-hpx_addr_t hpx_lco_and_new(intptr_t inputs);
+hpx_addr_t hpx_lco_and_new(int64_t inputs);
 
 /// Join an "and" LCO, triggering it (i.e. setting it) if appropriate.
 ///
